backend/requirements.txt
fastapi
uvicorn
sqlite-vss
httpx
orjson
aiosqlite
arq
numpy
locust
Faker
langdetect
pyjwt